# openapi.json
{"openapi":"3.1.0","info":{"title":"Meme Server API","summary":"Meme Server API for s3 storage","description":"\n    API сервера мемов.\n\n    ## Мемы\n\n    Можно:\n    * **Получить мемы** Meme List\n    * **Добавить мем** Meme Add\n    * **Получить мем по id** Meme By Id\n    * **Обновить мем по id** Meme Update By Id\n    * **Удалить мем по id** Meme Delete By Id\n\n    ","version":"0.0.1"},"paths":{"/memes/limit-offset":{"get":{"summary":"Meme List","description":"Листинг мемов с пагинацией","operationId":"meme_list_memes_limit_offset_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size limit","default":50,"title":"Limit"},"description":"Page size limit"},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Page offset","default":0,"title":"Offset"},"description":"Page offset"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LimitOffsetPage_MemePydantic_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/memes":{"get":{"summary":"Meme List","description":"Листинг мемов с пагинацией","operationId":"meme_list_memes_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_MemePydantic_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Meme Add","description":"Создать новый мем\n- **file**: изображение мема\n- **text**: текст мема","operationId":"meme_add_memes_post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_meme_add_memes_post"}}}},"responses":{"201":{"description":"Successful Response"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}},"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/memes/{id}":{"get":{"summary":"Meme By Id","description":"Получить мем по id\n- **id**: id мема в базе данных","operationId":"meme_by_id_memes__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meme_Imagebytes"}}}},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFound"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}},"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Meme Update By Id","description":"Обновить мем по id\n- **id**: id мема в базе данных\n- **new_file**: новое изображение мема\n- **text**: новый текст мема","operationId":"meme_update_by_id_memes__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_meme_update_by_id_memes__id__put"}}}},"responses":{"201":{"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFound"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}},"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Meme Delete By Id","description":"Удалить мем по id\n- **id**: id мема в базе данных","operationId":"meme_delete_by_id_memes__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFound"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}},"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_meme_add_memes_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"text":{"type":"string","maxLength":64,"minLength":1,"title":"Text","default":"Meme Description"}},"type":"object","required":["file"],"title":"Body_meme_add_memes_post"},"Body_meme_update_by_id_memes__id__put":{"properties":{"new_file":{"type":"string","format":"binary","title":"New File"},"text":{"type":"string","maxLength":64,"minLength":1,"title":"Text","default":"Meme Description"}},"type":"object","required":["new_file"],"title":"Body_meme_update_by_id_memes__id__put"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InternalServerError":{"properties":{"status":{"type":"integer","title":"Status","default":500},"detail":{"type":"string","title":"Detail"}},"type":"object","required":["detail"],"title":"InternalServerError"},"LimitOffsetPage_MemePydantic_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/MemePydantic"},"type":"array","title":"Items"},"total":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total"},"limit":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Limit"},"offset":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"LimitOffsetPage[MemePydantic]"},"MemePydantic":{"properties":{"id":{"type":"integer","title":"Id"},"text":{"type":"string","title":"Text"},"image_uuid":{"type":"string","title":"Image Uuid"}},"type":"object","required":["id","text","image_uuid"],"title":"MemePydantic"},"Meme_Imagebytes":{"properties":{"id":{"type":"integer","title":"Id"},"text":{"type":"string","title":"Text"},"image_uuid":{"type":"string","title":"Image Uuid"},"image_bytes":{"type":"string","title":"Image Bytes"}},"type":"object","required":["id","text","image_uuid","image_bytes"],"title":"Meme_Imagebytes"},"NotFound":{"properties":{"status":{"type":"integer","title":"Status","default":404},"detail":{"type":"string","title":"Detail"}},"type":"object","required":["detail"],"title":"NotFound"},"Page_MemePydantic_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/MemePydantic"},"type":"array","title":"Items"},"total":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total"},"page":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Page"},"size":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Size"},"pages":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Pages"}},"type":"object","required":["items","total","page","size"],"title":"Page[MemePydantic]"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}